+++ app/src/main/java/org/simple/clinic/patient/PatientRepository.kt
@@ -492,7 +492,7 @@ class PatientRepository @Inject constructor(
 
   fun clearPatientData(): Completable {
     return Completable
-        .fromCallable { database.clearPatientData() }
+        .fromCallable { database.clearAppData() }
         .andThen(reportsRepository.deleteReports())
   }