+++ app/src/sharedTest/java/org/simple/clinic/TestData.kt
@@ -55,6 +55,12 @@ import org.simple.clinic.protocol.sync.ProtocolPayload
 import org.simple.clinic.storage.Timestamps
 import org.simple.clinic.summary.teleconsultation.api.TeleconsultPhoneNumber
 import org.simple.clinic.summary.teleconsultation.api.TeleconsultationsResponse
+import org.simple.clinic.summary.teleconsultation.sync.MedicalOfficer
+import org.simple.clinic.summary.teleconsultation.sync.MedicalOfficerPayload
+import org.simple.clinic.summary.teleconsultation.sync.TeleconsultationFacilityInfo
+import org.simple.clinic.summary.teleconsultation.sync.TeleconsultationFacilityInfoPayload
+import org.simple.clinic.summary.teleconsultation.sync.TeleconsultationFacilityMedicalOfficersCrossRef
+import org.simple.clinic.summary.teleconsultation.sync.TeleconsultationFacilityWithMedicalOfficers
 import org.simple.clinic.user.LoggedInUserPayload
 import org.simple.clinic.user.OngoingLoginEntry
 import org.simple.clinic.user.OngoingRegistrationEntry
@@ -1062,4 +1068,84 @@ object TestData {
   ) = TeleconsultPhoneNumber(
       phoneNumber = phoneNumber
   )
+
+  fun medicalOfficer(
+      id: UUID = UUID.randomUUID(),
+      fullName: String = faker.name.name(),
+      phoneNumber: String = faker.phoneNumber.cellPhone()
+  ): MedicalOfficer {
+    return MedicalOfficer(
+        medicalOfficerId = id,
+        fullName = fullName,
+        phoneNumber = phoneNumber
+    )
+  }
+
+  fun teleconsultationFacilityInfo(
+      id: UUID = UUID.randomUUID(),
+      facilityId: UUID = UUID.randomUUID(),
+      createdAt: Instant = Instant.parse("2018-01-01T00:00:00Z"),
+      updatedAt: Instant = Instant.parse("2018-01-01T00:00:00Z"),
+      deletedAt: Instant? = null,
+      syncStatus: SyncStatus = randomOfEnum(SyncStatus::class)
+  ): TeleconsultationFacilityInfo {
+    return TeleconsultationFacilityInfo(
+        teleconsultationFacilityId = id,
+        facilityId = facilityId,
+        createdAt = createdAt,
+        updatedAt = updatedAt,
+        deletedAt = deletedAt,
+        syncStatus = syncStatus
+    )
+  }
+
+  fun teleconsultationFacilityWithMedicalOfficers(
+      teleconsultationFacilityInfo: TeleconsultationFacilityInfo = teleconsultationFacilityInfo(),
+      medicalOfficers: List<MedicalOfficer> = emptyList()
+  ): TeleconsultationFacilityWithMedicalOfficers {
+    return TeleconsultationFacilityWithMedicalOfficers(
+        teleconsultationFacilityInfo = teleconsultationFacilityInfo,
+        medicalOfficers = medicalOfficers
+    )
+  }
+
+  fun teleconsultationFacilityInfoMedicalOfficersCrossRef(
+      teleconsultationFacilityUuid: UUID = UUID.randomUUID(),
+      medicalOfficerUuid: UUID = UUID.randomUUID()
+  ): TeleconsultationFacilityMedicalOfficersCrossRef {
+    return TeleconsultationFacilityMedicalOfficersCrossRef(
+        teleconsultationFacilityId = teleconsultationFacilityUuid,
+        medicalOfficerId = medicalOfficerUuid
+    )
+  }
+
+  fun medicalOfficerPayload(
+      id: UUID = UUID.randomUUID(),
+      fullName: String = faker.name.name(),
+      phoneNumber: String = faker.phoneNumber.cellPhone()
+  ): MedicalOfficerPayload {
+    return MedicalOfficerPayload(
+        id = id,
+        fullName = fullName,
+        phoneNumber = phoneNumber
+    )
+  }
+
+  fun teleconsultationFacilityInfoPayload(
+      id: UUID = UUID.randomUUID(),
+      facilityId: UUID = UUID.randomUUID(),
+      medicalOfficersPayload: List<MedicalOfficerPayload> = emptyList(),
+      createdAt: Instant = Instant.parse("2018-01-01T00:00:00Z"),
+      updatedAt: Instant = Instant.parse("2018-01-01T00:00:00Z"),
+      deletedAt: Instant? = null,
+  ): TeleconsultationFacilityInfoPayload {
+    return TeleconsultationFacilityInfoPayload(
+        id = id,
+        facilityId = facilityId,
+        medicalOfficers = medicalOfficersPayload,
+        createdAt = createdAt,
+        updatedAt = updatedAt,
+        deletedAt = deletedAt
+    )
+  }
 }